Correct convert

+++ convert_png_to_xpm_and_del_png.sh
@@ -5,4 +5,9 @@ do
 	convert "$file" "${file%.png}.xpm"
 done
 
-rm *.png
+if [ $? -eq 0 ]; then
+	rm *.png
+	echo "\033[32mDone !\033[0m"
+else
+	echo "\033[31mConversion failed !\033[0m"
+fi
